+++ adapters/ciao.pl
@@ -1,7 +1,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %
-%  Adapter file for Ciao Prolog 1.19.0
-%  Last updated on January 2, 2021
+%  Adapter file for Ciao Prolog 1.20.0
+%  Last updated on March 31, 2021
 %
 %  This file is part of Logtalk <https://logtalk.org/>
 %  Copyright 1998-2021 Paulo Moura <pmoura@logtalk.org>
@@ -22,7 +22,7 @@
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
 
-:- use_package(iso).
+:- use_package(iso_strict).
 
 :- set_prolog_flag(multi_arity_warnings, off).