+++ RELEASE_NOTES.md
@@ -25,6 +25,12 @@ RELEASE NOTES
 3.46.0 - April ??, 2021
 =======================
 
+Prolog adapter and integration files
+------------------------------------
+
+* UPDATED: Require Ciao Prolog 1.20.0 or later version (due to backward
+incompatible changes in its libraries).
+
 Library
 -------